Put in skips in async control flow if no PDFs estimated
This commit puts in skips in the asynchronous control flow in walker's
Distributor and Integrator if there is no PDFs to be estimated at a given time
step. This happens when either the user did not request any PDF estimation in
the control file at all or the interval for estimating and outputing PDFs is
such that at the given time step PDFs are not estimated. (Estimation of PDFs,
especially multi-dimensional ones, can be relatively costly, and thus PDF
estimation is recommended only at certain intervals, i.e., not every time step.)

Since the statistics estimation has already been converted to use Charm++
reductions, testing without PDF estimation requested nicely triggers these skips
just put in and thus the effect of converting the statistics to reductions can
be measured. Here is some weak scaling data up to 24K cores:

The following setup dir1.q = ran on a single compute node:

===========================================
walker
  term  140.0   # Max time
  dt    0.05    # Time step size
  npar  1000000   # Number of particles
  ttyi  100    # TTY output interval

  rngs
    mkl_mrg32k3a seed 0 end
  end

  dirichlet     # Select Dirichlet SDE
    depvar y
    init zero
    coeff const
    ncomp 2  # = K = N-1
    b     0.1    1.5 end
    S     0.625  0.4 end
    kappa 0.0125 0.3 end
    rng mkl_mrg32k3a
  end

  statistics
    <Y1>
    <Y2>
    <y1y1>
    <y2y2>
    <y1y2>
  end
end
===========================================

The above setup then was customized to enable weak-scaling by varying the
number of particles in the ensemble:

$ grep npar dir* | view -
dir1000.q:  npar  1000000000   # Number of particles
 dir100.q:  npar  100000000   # Number of particles
  dir10.q:  npar  10000000   # Number of particles
   dir1.q:  npar  1000000   # Number of particles
 dir500.q:  npar  500000000   # Number of particles
  dir50.q:  npar  50000000   # Number of particles

All runs were carried out with a virtualization parameter u = 0.9, yielding
roughly 10 work units per CPU. Compared to a classical (pure-MPI) strategy which
would correspond to no virtualization (u = 0.0), this is a tougher problem,
since potentially more communication is involved as the work units are only the
10th in size of what they would be with u = 0.0.

Timings (in min:sec) measuring the elapsed wall-clock going from t=900 to t=1000
are as follows.

nn(1)      nc(2)     nored(3)     red(4)  red+skippdf(5)
    1         24         0:04       0:19            0:09
   10        240         1:13       0:26            0:09
   50       1200         1:23       0:41            0:04
  100       2400         3:23       1:54            0:04
  500      12000     canceled   overload            0:06
 1000      24000 only 4 steps   overload            0:11

The first column is the number of compute nodes, each with 24 cores (second
column).

The third column uses no Charm++ reductions at all and has no skips for
statistics or PDFs, i.e., corresponds to commit 47f936e. As the data shows, that
approach does not scale at all and the messaging system eventually gets
overloaded with a lot of small messages at high core counts.

The fourth column contains Charm++ reductions for the statistics, but does not
yet have the skips for PDFs which then introduce unnecessary syncronization
points in the control flow even if there are no PDFs estimated as is the case
here. This setup corresponds to commit 0c1ca13. The data shows that while the
scaling improved a lot compared to the 3rd column, not skipping the PDFs (which
in that commit has not yet been implemented) still takes an unnecessary toll on
the scaling and the messaging system. The cases with 500 and 1000 nodes
overloaded the messaging system.

The fifth column shows the same weak scaling experiment but now Charm++
reductions for the statistics estimation AND the skips in place for if there is
no PDFs are estimated (which is the case here). The fifth column shows
near-perfect weak scaling up to 24K compute cores. This is, of course, also the
case if there are no statistics estimated at all (measured but not discussed
here) which is an easier case. The skips put in place for if there is statistics
estimation by commit 9cd1b4f.

Also there is a couple of interesting facts to note about the runs in the fifth
column:

1. The computational load and its corresponding memory usage for these runs can
be approximated by only considering the particle array, which takes
1000000*8/1024/1024/1024/64*100 = 0.011642% of the available (64GB) memory per
compute node. This is miniscule for the machine. In other words, the scenario
tested operates on a very tiny data set, thus the amount of
computation/communication ratio is very small. A more realistic case therefore
should not scale worse.

2. Since the data set is very small, for the cases that use 50, 100, and 500
compute nodes (1200, 2400, and 12000 CPU cores, respectively), the data fits
into the cache better than the other runs, and thus yields superlinear speed-up.

In summary, this strategy promises to be very performant, since it allows very
good weak scaling and overlap of computation and communication.

Next is to implement Charm++ reductions for the PDF estimation as well and
measure if degrades this performance in any way.