Adjust 100% of code coverage
This is so that the 100% is more accurately computed for the unit and regression
suites.

Since unittest is not a regression test (and is not run as a regression
test), it should not be built and counted to the 100% of teh code that is
supposed to be tested by the regression suite.

Also, regtest is soon to be removed, so it is now already taken out of the unit
test and regression test coverage report.

This this change include:
 - removing regtest and unittest from the regression suite, and
 - removing regtest from the unit test suite