Adjust update_doc.sh for more accurate coverage
Revamped the update_doc.sh script, which now wipes the work directory between
the generation of unit test coverage, regression test coverage, and doc
generation.

Also, the script is now working in its own work directory and cleanes up after
itself. Thus it is more independent of the build dir that it used to work in.