Mesh partition in inciter now starts with parallel read from file
This codes is still a work in progress. It works and introduces Spawner, a
Charm++ group facilitating tree-based scalable Charm++ array creation of
Performers. This commits converts the initial mesh partitioning to start from a
parallel and thus distributed read of the mesh from file and is made to work
with the initial distributed mesh. Zoltan partitions in parallel, produces its
result in parallel in a distributed fashion. This allows taking g_element,
holding the mesh element indices and associated chare ids, out of Charm++'s
global-scope distribution of data.

While this setup seems to work fine, Zoltan produces awful partitions, because
some mesh nodes are non-unique in the aggregated mesh as a result of handling
the mesh in a distributed fashion. After talking to Karen Devine at SNL, project
lead of Zoltan, also attempted building the distributed hypergraphs for Zoltan
based on element connectivity. While that would be the correct thing to do,
instead of the current mesh-node-based hypergraphs since the field solver is
element based, building element connectivity on a distributed array does not
yield the full graph as some elements can have no neighbors at all and certainly
less than their true neighbors compared to the complete, non-distributed mesh.
I don't see a straightforward way to fix this easily and now decided to try
Zoltan2, which has a mesh adaptor that, according to Karen, will take care of
this for me: I only have to pass the distributed mesh and Zoltan takes it from
there, builds the hypergraphs, partitions in parallel, etc. That is our next
step. For more documentation here is an excerpt from our email exchange with
Karen:

===
Jozsef:

I think it would be a good idea to convert to mesh-based load balancing, and I'm
almost done with that, but I just thought of a potential problem this morning
that may be a problem independent of whether node-, or element-based
partitioning one is doing. If I read the mesh in parallel, it is inevitable that
I will not have the complete graph even in aggregate fashion across all MPI
ranks. Since I only read part of the mesh, some cells on a rank will not know
about one or some of their neighbors that happen to be on other ranks. In fact
some cells will have no neighbors at all, which is certainly not true if one
looks at the aggregated mesh. Will this be a problem for Zoltan? I assume it
will be, because the hypergraphs I compute on the partial meshes on each rank
will not contain all edges of the complete graph/mesh. So with my previous
approach, I have duplicated nodes, since I read elements which may connect to
elements on other ranks, and converting to element-based hypergraphs I don't
give Zoltan the full picture. It seems like all this trouble is due to the fact
that I'm reading the mesh in a distributed fashion.

Now I'm thinking that my approach should be something like the following. I
should probably read the full mesh on a single rank, compute a full hypergraph,
and distribute parts OF THE HYPERGRAPH instead of parts of the mesh (which now
have full connectivity information) to other ranks.  Pass those to Zoltan, which
gives me back distributed partition IDs assigned to mesh elements and I take
that from there.

I'm still not very happy with this because I need to read the mesh on a single
rank and need to compute the full hypergraph on a single rank. What if that does
not fit into memory eventually?

===
Karen:

Would you be interested in trying Zoltan2 instead of Zoltan?  I ask because
Zoltan2 can resolve the +concern you have about the distributed mesh input.

Zoltan2's interface includes a MeshAdapter.  Instead of making the user (you)
build a graph or hypergraph, +Zoltan2 lets the user just describe the mesh.  So,
for your application, you could read chunks of the +mesh into processors, and
let each processor describe its own chunk; you wouldn't have to worry about the
+off-processor chunks.

Then, to do graph or hypergraph partitioning, Zoltan2 can construct the graph or
hypergraph based on the +mesh adjacencies (e.g., element to vertex).  Here's a
rough outline of how Zoltan2's logic could work:

-  You read your mesh into processors so that each processor has a chunk of
   elements
-  You provide an implementation of Zoltan2's MeshAdapter describing your chunk
   of elements, including, +for example, the local elements' global names and
the mesh nodes to which the local elements are +connected.
-  Zoltan2 can then derive graph/hypergraph edges, even across processors, from
   the local elements and +their nodes.  For example, Zoltan2 can create a graph
whose edges connect elements that share nodes.

Zoltan2 is newer code, so you might need to work with us a little to resolve
issues, but we are motivated +to respond quickly.  A project at Sandia is
adopting Zoltan2's MeshAdapter as well, so we are actively +supporting and
enhancing that part of the code.

Zoltan2 requires a C++11 compiler and builds with Trilinos.  I'd be happy to
help with that.  We could +work via phone/email, or we can arrange a day visit
to hack it out together.
===

More details for this commit:

 * ExodusIIMeshReader can now return the number of elements for a block, i.e.,
   elemtype

 * Charm++ global-scope broadcast now does not have to see g_element, storing
   the element ids for chares as it comes out of Zoltan distributed

 * Fix screen output of mesh stats, in sync with the parallel read and the
   result of the distributed mesh in memory across MPI ranks

 * g_element is now a map associating mesh element ids to chare ids. As this is
   now distributed, not all chare ids are present on all ranks (obiously) and
   chare ids must be known to arbitrate proper communication to fellow PEs

 * UnsMesh graphsize() now returns not the connectivity divided by the number of
   nodes in elements, but equates the graphsize with the number of unique node
   ids in the connectivity