Collect LinSysMerger stats only from rows
Previously, we have collected statistics of communicated vs communicated + owned
matrix nonzeros and right hand sides. This contains approximately the same
information so now we only collect the number of communicated vs communicated +
owned rows and collect this information when the rows are contributed. This has
the advantage that it only happens once during setup of LinSysMerger and not
during time stepping - provided the mesh does not change as is the case at this
time.

Also removed some debug printouts and augmented timing infos with number of PEs
information.