ChareID-elemID map distribution seems to work
This is the first step in putting in the missing link for a reasonable linear
system assembly. The first step (this commit) is to take the global mesh element
IDs associated to chare IDs by the partitioner and distribute them to their
owner PEs. The next is to compute a reordering, i.e., a permutation, of the
global mesh node IDs so that all chares working on a PE predominantly contribute
to the chunk of the liner system owned by the PE. In other words, the reordering
computes a permutation of the global mesh node IDs read from the file
(inefficient) to a new ordering needed by a more efficient linear system assmbly
and solution.