Reordering seems to work
This is the second step in putting in the missing link for a reasonable linear
system assembly. This commit puts in the lamest style reordering that can be
possibly imagined: it computes the permutation on a distributed array in a
serial fashion with increasing cost of communication as the work is done by PEs
with successive IDs. This demonstrably does not scale at all, its cost is
super-linearly increasing with the number of PEs as well as the number of mesh
cells. In other words, this badly sucks at this point, but seems to work, which
is the first step towards a reasonable linear system assembly.

Note that though the cost of the distributed reordering is superlinear and does
not scale, still this needs to be done only once during setup and we
(supposedly) avoid a very non-optimal communication pattern during the linear
system assembly in LinSysMerger, which must happen every time step.

Initial measurements of the communication cost in LinSysMerger indicate that the
communication cost has somewhat decreased with this reordering in place, but
there seems to still be yet another step required to make this reordering matter
more, and as a result, decreasing the communication cost further. That is: we
probably should relax the equal number of rows on all PEs for the linear system.
This is currently computed by simply assigning the same number of rows to each
PE, but the communication cost could possibly be reduced with an unequal number
of rows assembled by each PE - for the cost of each PE having to assemble an
unequal number of rows, which may yield a load imbalance across assembly, and
the solution of the linear system during the Hypre call.

A simple strategy yielding an unequal distribution of the number of rows is to
compute the lower and upper row indices based on (during) the reordering so that
more of the node IDs assigned to a PE fall into that PEs lower-upper row ID
band. But other strategies can be cooked up. In the end we somehow need to
minimize communication during the linear system assembly.