Unequal number of rows per PE for LinSysMerger
This is the final missing piece for a reasonable linear system assembly. This
commit relaxes the equal number of rows on all PEs for the linear system and
instead adjusts each PE so that the communication costs among PEs are minimized.
We simply find the largest node ID assigned on each PE by the reordering and use
that as the upper global row index. While this rarely results in equal number of
rows assigned to PEs, potentially resulting in some load-imbalance, it yields a
pretty good division reducing communication costs during the assembly of the
linear system, which is more important than a slight (FLOP) load imbalance. We
also pre-compute the communication cost as a diagnostic so we don't waste time
during linear system assembly. The cost is computed in parallel and
asynchronously.

Additional optimizations are highly likely possible. For example, switching from
maps to unordered maps.