Distributed mesh node reordering seems to work
This commit puts in an infintely better mesh node ID reordering than the
previous one, which was effectively serial (each PE called PE+1 when it was done
with its reordering), in addition each PE was practically snowballing an
ever-larger map associating old to new ids in order to find mesh connections
between PEs, which needed to be passed downstream to all PEs. This was wasteful
and very-very inefficient. For example, on a few million-cell mesh and 200-300
PEs, the reordering easily took several (5-6) minutes, compared to a time step
later, which was maybe a second.

Background: This reordering is needed because the mesh node ordering read from
file is suboptimal compared to what we need for the matrix and vector assembly
and the linear solve during time stepping. Thus we neeed a reordering, i.e., a
permutation, of the global mesh node IDs so that all chares working on a PE
predominantly contribute to the chunk of the liner system owned by the PE. In
other words, the reordering computes a permutation of the global mesh node IDs
read from the file (inefficient) to a new ordering needed by a more efficient
linear system assmbly and solution.

The new algorithm precomputes the communication maps for each PE and does the
reordering largely in parallel. At a high level, it works as follows.

Each PE works on a chunk of the mesh that was assigned to it by the new ordering
from the partitioner. Each PE has two groups of nodes to reorder: (1) those that
can be reordered without any interaction from other PEs, and (2) those that
cannot be, because other PEs assign IDs for them whose PE indices are less than
the PE. In other words, the reordering (as in the old algorithm) starts on PE 0,
assigning new node IDs, without any communication necessary from any other PE.
Then PE 1 only reorders those nodes that can be reordered without affecting PE
0, because they don't exist on PE 0. Next is PE 2, which needs to know which
nodes are assigned new IDs by potentially both PE 0 and PE 1 if the mesh on PE 2
connects to the mesh on PE 0 and/or PE 1, and so on. The last PE, thus needs to
know about potentially all other PEs reodering work, i.e., which nodes are
reordered on them if the last PE connects to any or some of the mesh held on PEs
with indices less than the last one.

The new algorithm this commit puts in takes about 2-3 seconds for the example
mentioned above, compared to 5-6 minutes. This includes both the initial
construction of the communication maps for all PEs and the distributed
reordering including the communication necessary to match the node IDs among PEs
that share nodes.

New timings indicate that the bottleneck now is reading of the owned mesh graph
on each PE. This not terrible, a few seconds for an example mentioned, but
demonstrably I/O-bound and its performance deteriorates as more PEs try to
access the same, especially, if the reading requires random seeks, as need.

A better way of doing this is to communicate the graph connectivity among the
PEs: after reading contiguous chunks and partitioning, we have the new owners
assigned. At this point we need to send the mesh connectivity to the correct PEs
instead of re-reading them in an unordered fashion from file. That's next.