Add mix mass-fraction beta SDE homogneous decay coefficients policy with DNS hydrodynamics time scale
This commit adds a new coefficients policy, i.e., model, for the
mix-mass-fraction beta SDE in walker, that enables feeding in an arbitrary
function evaluated at a time as the hydrodynamics time scale. We use this to
test the behavior of the variable-density material mix-model, under development,
independent of hydrodynamics.