Pass the physical time down to diffeq::advance()
This was needed because the previous solution, i.e., not passing down the
physical time to diffeq::advance(), incorrectly computed the physical time for
MixMassFracBetaCoeffHydroTimeScaleHomDecay, the only coefficients policy so far
that needs the physical time. Interestingly, the problem only came out with
virtualization larger than 0 and screwed up the solution. This commit fixes that
issue.