Performer now stores the lhs sparse matrix as std::vector
Instead of std::map. Using associative containers has been a nice and safe
vehicle during development, and now we have a good enough understanding to
change the storage to std::vector. The performance improvement is roughly 2-2.5x
for the matrix assembly, in LinSysMerger::charelhs() and LinSysMerger::addlhs().
Filling of the matrix, in Performer::rhs(), is unmeasurable (or there is a bug
in the asynchronous collection of timing statistics - unlikely, because other
timings seem reasonable).