Move creating workers from Conductor to Partitioner in Inciter
Background: Conductor is a single Charm++ chare, while Partitioner is a Charm++
chare group and thus there is one (and only one) Partitioner object guaranteed
to exist on all PEs. Any work done by Conductor is practically serial - as can
be seen by Charm++'s performance analysis tool, Projections.

Reason for this change: Using Projections, identified a large amount of (serial)
work being done after the distributed mesh reordering algorithm is finished by
Partitioner. Though this is only done once during setup, it can still take up a
significant amount of time, delaying the start of time stepping. The reason for
this (serial) work, that so far was started off in Conductor::prepared(), was
two-fold: (1) estimate the communication cost of the merging the linear system
later (mere diagnostics), and (2) initialize Performer Charm++ chare array
elements by sending each its global mesh element connectivity, i.e., node IDs,
it contributes to and the old->new mesh node ID map (after reordering). While
(1) is only a diagnostic (still it should not put in an unnecessary barrier, as
it has done so far), (2) is absolutely necessary for each Performer to start
work, preparing to solve PDEs and integrate in time.

As it turns out, neither (1) nor (2) must be done by Conductor and can be done
by Partitioner, i.e., in parallel instead of in serial. This was facilitated by
merging the functionality that has been done by Spawner, another Charm++ chare
group beside Partitioner, which was to be used to spawn finer-grained work done
by Performers (potentially many more per PE). Also, the original purpose of
Spawner was to have fine control on the creation of Performer chare array
elements, i.e., passing a specific chunk of the mesh element connectivity and
node IDs to them via their constructor. This is (as has been) done via the
insert() member function, which allows fine control on array element placement
and constructor arguments. Thus Spawner is really not required and so its
functionality has now been incorporated into Partitioner.

As a result of the above refactoring, not only the serial section following the
mesh-ordering has been eliminated, but during time stepping a much tighter CPU
utilization is achieved, i.e., the idle time has been significantly reduced,
compared to the previous commit. This nice side effect of our main purpose is
most likely due to the fact that there are no separate Performer chare arrays
held by a Charm++ chare group elements, but instead a single array is held by
Conductor, and thus most likely, the runtime system can do a much better job at
scheduling a single array and its elements. Client code in LinSysMerger, as well
as Performer, also simplifies, since we don't have keep accounting for different
worker proxies, as there is only one, and do not have to account for different
chare array element IDs (local vs global).

Note that there is a still a somewhat significant serial portion in the setup:
this is before the distributed mesh node ID reordering starts, computing the
communication map necessary to do the reordering in parallel. This shows up in
Projections as Conductor::addNodes(), computing the communication map for each
PE as information becomes available for each.

Also, now a lurge chunk of the CPU time during time stepping is spent on
computing the right hand side, which what is expected for nature of the this
work.

More specific details for this commit:

 * Add Print::diag(), which allows a simple diagnostics output to screen,
   instead of having to call diagstart() for which there would be no diagend().

 * Spawner has now been removed, as discussed above.

 * All the Charm++ proxies for Performer, LinSysMerger, and Partitioner, are now
   held by Conductor, instead of Spawner (or Partitioner) holding Performer.

 * Conductor::prepared() is gone. Its work has now been eliminated since the
   connectivity and chare ID map it was receiving are computed and available in
   Partitioner, ready to be used (by Partitioner and Performer objects).

 * Conductor::computeCost() is gone. Its work is done by Partitioner::bounds(),
   and Partitioner::lower().

 * Conductor::query(), Conductor::costed(), and Conductor::estimate() are gone.
   Their work is done by Partitioner objects, collecting the average and
   standard deviations of the communication cost of merging the linear system,
   by (asynchronously) reducing the results to Conductor::aveCost() and
   Conductor::stdCost(). These are standard Charm++ typed reductions.

 * Creating Performer chares are now done by Partitioner objects (one each PE),
   instead of by Conductor. That's a lot more scalable than the approach before
   (while the Performer proxy is now held by Conductor which simplifies code as
   discussed above).

 * This refactoring also frees up some memory in Conductor's state:
   m_connectivity, m_chcid, m_div, and due to removing explicit timing code
   before, m_arrTimestamp and m_grpTimestamp.

 * LinSysMerger now computes the inverse of the PE-division map (computed by
   Partitioner) instead of simply receives it. This results in less data being
   communicated so it should be better performance though I have not measured
   this specifically. This is done by broadcasting the bounds from each
   Partitioner objects, i.e., all PEs, to all LinSysMerger objects, i.e., to
   all PEs. The broadcast is initiated at the end of Partitioner::create(),
   because once the inverse division map is complete on all PEs, LinSysMerger
   instructs Conductor to continue with the setup of Performers, see
   Conductor::setup(), and LinSysMerger::charerow(), called by
   Performer::setup() needs the inverse maps. Communicating the bounds to
   LinSysMergers is an all-to-all, but only a couple of integers are transmitted.
   A copy of the inverse map on each LinSysMerger PEs is necessary because each
   needs to know which PE a given global node (or row) ID belongs to (so that
   each PE knows where to send data to that it receives from its workers).

 * LinSysMerger now only has to store a single worker proxy, which reduces
   its state (m_worker is not necessary) and simplifies its logic. (Note that
   this is how it used to be originally.)