Update regression tests to use RNGs that support leapfrogging
Some MKL RNGs, used previously by some of the tests, do not support leap
frogging, so in parallel they have generated the same exact series of random
numbers. As the previous commit puts in a throw if those RNGs are attempted to
be used in parallel, these tests would fail. This commit switches the those RNGs
to MCG59 which supports leapfrogging.