Fix walker's time stepping logic
The reason for work that led to this commit was to put in estimation of
statistics before the first time step is taken in walker. Previously, after
initialization, we started advancing the stochastic equations in time. However,
some equations, e.g., the one we are currently working on, need meaningful
statistics, estimated from the initial conditions before meaningful coefficients
can be computed. This has not been possible previously, because after setting
the ICs (on the samples) we started advancing the time step.

This commit puts in logic that ensures that in the first "time" step we only
estimate statistics (and potentially PDFs if requested), without attempting to
advance the equations. This is so that the next "real" time step, when we will
advance the equations in time, we have good statistics from ICs. In other words,
iteration 0->1 is estimating statistics from the ICs, and iteration 1->2 advance
in time. Note that while we do not advance the equation in time during it 0->1,
the pyhsical time (t) does increase by dt.

Additionally, we have refactored the time step logic for walker, that is now
similar to that of inciter's. As a result, many a of baselines have changed in
the regression suite - modulo statistical variability.

Also removed timing of the initialization step, which removes an unnecessary
reduction.