Make building of the communication maps parallel
Building of the communication maps is a prerequisite of distributed mesh node ID
reordering during inciter's setup. Previously this step was done by a reduction
from Partitioner to Conductor, each Partitioner PEs sending their unordered
global mesh node IDs to Conductor, which built a vector of hash-set for each PE,
which then were used to build the communication maps for each PE in serial (by
Conductor.

This commit removes the reduction and makes each Partitioner PE build its own
vector of hash-sets for the PE it needs to interrogate in order to build its
communication map. This enables each Partitioner chare to compute its own
communication map and removes multiple synchronization points with Conductor.
This also has the advantage of the communication maps being stored by
Partitioner, i.e., a part on each PE, in stead of by Conductor for all PEs.

Once the communication maps are done, the final step before reordering can start
is to compute the global mesh node ID offsets for each PE. This is now also done
in parallel, among fellow Partitioners.

The result of this change set is that now the complete setup before distributed
mesh node ID reordering is done in parallel in a distributed fashion and
asynchronously, with significant overlap of communication and computation,
evidenced by Charm++'s Projection tool as well as the large performance
improvement: taking 10 time steps using a 76M-cell mesh on 64 CPU cores is
reduced from 2:03 (2 minutes 3 seconds) to 47 seconds, a 2.6x speedup, and the
total time from 5:45 to 2:15. Thus the setup for mesh reordering (which is what
this commit was mostly intended to improve) goes from 3:42 to 1:28, a 2.5x
speedup.

More timing info (in mins:sec) for the same 76M mesh taking 10 time steps.

nPEs    timesteps            total_runtime
 64     was:2:03 now:0:47    was:05:45 now:2:15
128     was:1:01 now:0:20    was:06:55 now:1:22
256     was:0:30 now:0:09    was:11:26 now:0:35

Which means that now not only the time stepping exhibits perfect strong scaling,
but the total run, i.e., the initial setup as well.

It is not clear why there is a large improvement in the time stepping phase, but
I'll take it. One hypothesis is that this commit removes Conductor's SDAG code
entirely, which probably removes a lot of complexity from associated Charm++
code for that class.

Oh, and this commit also fixes the nasty bug introduced in 14ada79, which in
Partitioner.h replaced

return !std::any_of( cbegin(comm), cend(comm),

by

return !std::any_of( comm.cbegin(), comm.cbegin(),