Revamp computing the communication map before node reordering
This commit improves the algorithm that computes the communication maps before
distributed parallel global mesh node ID ordering takes place in Partitioner for
Inciter. Previously, this step was already parallel, the performance was okay,
but now it is great with very tight CPU utilization, as measured by Charm++'s
Projections.

Background: The communication map associates, on each PE, a list of global mesh
node IDs to PEs, whose indices are lower than PE. These are node IDs the PE
needs but does not assign new mesh node IDs during reordering. For these nodes
the new IDs are assigned by lower PEs and thus PE needs to receive these new
IDs. In order to know which these node IDs are and what PE assigns new IDs to,
we build this communication map before reordering.

Previously, the communication maps were computed by each PE sending its node IDs
it has read from file to all PEs upstream. This was so that upstream PEs can
search these IDs in order to find all node IDs they will need to receive (and
therefore not assign) during reordering. However, sending all this data was
suboptimal for two reasons: (1) A lot of data had to be sent, (these are all
domain mesh node IDs a PE has read from file and will contribute to during
calculation), even those that will not be needed for building the communication
maps; and (2) the load was imbalanced, since lower PEs needed to send an
inproportionally large amount of data to all higher PEs compared to higher PEs.

A better solution, put in by this commit, is to send the request to do search to
a PE and get a result back during building the communication maps. This involves
a lot less data movement and the load is also more balanced. At a high level,
the idea is as follows. Each PE sends a query to all other PEs along with a
vector of its node IDs. Upon receiving the query, each PE looks up all the node
IDs and generates a mask that indicates whether the node was found on the PE or
not. The length of the mask is the same as the length of the incoming node ID
vector and its values are booleans (implemented as integers, since that is less
error prone than vector of booleans). The resulting mask is then sent back to
the querying PE, which then collects them and generates its communication map.

This new algorithm balances the load in a way that while lower PEs will receive
a lot more requests than higher PEs, higher PEs will receive a lot more answers
(masks) than lower PEs. Example: PE 0 will have to do the most lookups and needs
no masks, while PE N will do the most work receiving the result of lookups but
does not need to do queries. Though these two would be imbalanced by themselves,
together they nicely overlap with very tight CPU utilization extremely small
idle time. This is implicit evidence that the queries and the processing of the
results of the queries are approximately equal load.  This is compared to the
previous solution, in which higher PEs needed to receive more and more data
while the lower PEs had nothing to do. In addition, in the new algorithm the
queries are done in a broadcast fashion, even though this would not be strictly
necessary to perform the task at hand, since only higher PEs need to query lower
PEs. However, using broadcast performs much better than individual sends and
thus there is logic put in so that the results are simply ignored if they are
not needed for computing the communication maps on a PE. More on this is
detailed in the code as comments.

The performance is tested using a 3.5M-cell mesh taking 10 time steps on 36
cores, as well as using a 76M-cell mesh on 64, 128, 256, 512, 1024, and 4096 CPU
cores.  We still have strong scaling for not only time stepping but also for
setup (which this commit mainly affects and which slightly got shorter). More
details on performance timing:

mesh    CPUs    virt    10steps    total
-----------------------------------------------------
76M      64     0       0:47       was:2:15, now:2:15
76M     128     0       0:20       was:1:22, now:0:57
76M     256     0       0:09       was:0:35, now:0:31
76M     512     0       0:04       was:0:23, now:0:20
76M    1024     0       0:02       was:0:23, now:0:15
76M    2048     0       0:01                     0:16
76M    4096     0       0:00                     0:28

The setup is now in a very good shape with a potential to scale to a very large
number of cores. This includes the lowest idle-time ever during setup. There may
be some room for improvement in the reordering step which still shows some idle
time.