Increase abstraction of particle data layout
This commit increases the data layout for particle data to a higher level. The
functionality remains the same and now data stored at mesh nodes (in a
mesh-based solver) can also use the two different compile-time configurations
for data layouts as the particle data could so far.

Also reworded and augmented the discussion on the data layout in the design
document doc/pages/layout.dox.

This is the first step in introducing various physics in Inciter in a
configurable fashion.