Add gnu compiler to drone build excluding gnu/libc++
Note that this commit aguments the drone build matrix with the 'exclude'
directive, which makes drone skip some entries of the build matrix.

Here we instruct drone to generate a 24-entry build matrix (3 compilers, 2 build
types, mkl or not, and two implementations of the C++ standard library, which
yields 24 different builds). Using the new 'exclude' keyword we instruct drone
to exclude all gnu/libc++ combinations, since there is not an easy way to use
gcc with a standard C++ library other than libstdc++. This correctly excludes 4
builds.

Note that the 'exclude' keyword is not yet part of the upstream drone
distribution, only if/when drone/drone#1514 is merged.