Replace _ with . in inciter's output filename
ParaView now automatically recognizes multiple files as "<filename>.<PEid>" as
multiple exodus files. This is compared to the previous behavior, where if the
filenames were "<filename>_<PEid>", they were not detected as Exodus files but
could be still read in correctly.

Also replace constructing the filename with a single string constructed at once.
Opt for safety here, requiring adding to Performer's state, but correct code is
more important than performance (that is so little probably cannot even be
measured).