Function objects args, bound via bind, now lvalues
Arguments to function objects, bound by boost::bind and stored in
boost::factory, have always been required to be lvalues. See the documentation
on boost::factory.

However, Boost 1.60 now enforces this explicitly. (It appears that this version
of the boost::factory/bind operation chain now has some optimizations which
explictly requires this.) As a result, the DiffEq and PDE factories can now only
store std::functions that hold constructors bound to lvalues and lvalue
references, and therefore the signature of the std::function must be const ref
or ref and pass-by-value arguments are not allowed. Thus the type of the
constructor signature stored in the std::function object held by the factories
were changed. Note that this does not affect the model constructors, i.e., the
"derived-class" constructor signatures, e.g., Dirichlet, AdvDiff, etc., which
still take arguments by value. This is facilitated by changing the host
constructors to take forwarding references (which now also do perfect
forwarding) of the model constructor arguments, instead of the previous solution
which created temporary copies of the arguments passed to the model
constructors, see e.g., DiffEq.h.