Make clang not to resolve symbols when creating a shared library on Mac OS
Resolving symbols only when linking executables and not when creating shared
libraries is consistent with gcc, intel, and clang on linux and gcc on mac.

Also put in some includes in the diff eqs to make apple/clang find
tk::particles.

Also enable static link on mac. We might be able to support this though this
does not work yet.