Make clang on mac os find libc++
Interestingly, running the TPL cmake on mac os behaved so far differently
depending on the presence of cmake cache. If the cache was empty, libc++ was
not found, if the cache was not empty, libc++ was found.

As it turns out, the BUILD_SHARED_LIBS variable (set in cmake/BuildShared.cmake)
was uninitialized in the tpl CMakeLists.txt, and thus libc++'s find_package
defaulted to looking for static libraries only and so it was not found. However,
when the cmake cache existed BUILD_SHARED_LIBS was already initialized and so in
the second run it was looking for the dynamic library.

This somewhat convoluted behavior is now fixed by setting BUILD_SHARED_LIBS
before looking for libc++, which now does the same thing independent of the
cache.

Additionally, we now disallow static libraries on mac os if the compiler is not
gnu. Previously, this was allowed in the hope that this might work, but
disallowing it seems like a better choice, since this will make the behavior
consistent on both linux and mac os: static libs (and static linking) is only
supported with the gnu compilers all platforms.