Charm++ can now be built on 32-bit OS and w/o mpi wrappers
This commit generalizes the cmake logic for building Charm++.

1. It allows not loading any modules, i.e., not setting any environment
variable, and not passing any compiler settings to cmake, letting cmake find a
system-default compiler and find a system-wide mpi, i.e., use whatever is
default native on a system.

This required makig sure we can not only build using the mpi wrappers but the
underlying compilers as well. This is simply done by not specifying 'mpicxx',
but specifying nothing, as the Charm++ compiler. Note that '' for the Charm++
compiler was required for Mac OS and this behavior is kept. For the other TPLs
we still use the mpi wrappers, which are found by cmake's FindMPI.cmake.

2. Now we can also build Charm++ on a 32-bit OS. This is currently only tested
up to the TPL-build, which successfully builds all TPLs, except Trilinos, in
which Kokkos does not (yet) support a 32-bit build (verified by Christian
Trott). This has currently only been tested on linux (not on mac) and without
the mpi wrappers.