Make LDFLAGS and CPPFLAGS work with multiple dirs
These are used for the netcdf TPL build. If there are multiple -L and -I
arguments, the cmake external_project_add) string trickery have failed and have
not been exercised so far. Now thanks to the new TPL combination build matrix,
this problem was exposed and now (hopefully) is fixed.