Check if the 'head_ref' file exists when detecting the git sha1
This is necessary to be able to use TeamCity's "VCS checkout mode" as
"Automatically on agent (if supported by VCS roots)" setting. Though TeamCity in
this mode does a checkout on the build agent thus git metadata exists, still
.git/refs/heads/ is empty, so the file 'head_ref', created in the
get_git_head_revision() cmake function, does not exist, which causes the cmake
function get_git_head_revision() to fail, breaking the build. This commit wraps
attempting to read from the non-existent file with a test for existence, so the
empty .git/refs/heads directory does not break the build.