Fall onto 'git describe' if hash not found by get_git_head_revision()
This way 'git describe' can still find the commit sha1. Since TeamCity's
agent-chackout does not populate the .git/refs/heads/ directory,
get_git_head_revision() does not find the git sha1. This commit makes sure that
even in that case we find the commit sha1 by executing 'git describe'. This way
the TeamCity builds will have the commit sha in the logs as well as in the
generated executables.