Fix a weird "lazy symbol binding failed" error on Mac OS
Here are some of the runtime errors only appeared on Mac OS with inciter:

dyld: lazy symbol binding failed: Symbol not found: __ZN2tk12LinSysMergerIN7inciter16CProxy_ConductorENS1_16CProxy_PerformerEE8wait4solEv
  Referenced from: <quinoa>/build/Inciter/libInciter.dylib
  Expected in: flat namespace

dyld: lazy symbol binding failed: Symbol not found: _HYPRE_IJMatrixCreate
  Referenced from: <quinoa>/build/clang/LinSys/libLinSys.dylib
  Expected in: flat namespace

Though I'm not 100% sure how to explain why the fix works, it has two
ingredients, which this commit puts in:

(1) Fixed building our own hypre on Mac, producing a dynamic library, instead of
using the macports-installed one, which only provides a static library.

(2) We needed to put in "extern template ..." to both Conductor.C and
Performer.C. These two source files are part of libInciter which was missing the
LinSysMerger template.

A potential explanation is that the linker on Mac OS instantiated the
LinSysMerger template multiple times so inciter at runtime got confused which
one to use, or the linker did not instantiate it in a way that would be
available for libInciter. Interestingly, fiddling with the link order, linking
the final inciter executable, did not help.

Thanks for Francisco @franjgonzalez for finding the bug.

Also removed the recommendation of installing hypre using macports on Mac OS
from the README, which will trigger building our own.