Put in diagnostics collection logic to inciter
For now, this only works for a single compns, as the diag vector size (5) is
hardcoded in Conductor's ctor. This commite makes progress on the same work
started by and being discussed in 8a72129.

The logic to optionally collect diagnostics from Performer chares to host
Collector is now in place. This is done via adding SDAG logic into Conductor as
outputing the one-liner report can only be done if both the diagnostics have
been contributed as well as the evaluation of the time step is complete. If
diagnostics are not collected in a given time step, only the latter needs to be
finished, in which case diagnostics are skipped.

Interestingly, Charm++ reduction targets are required to be non-const memberg
functions.

New Conductor SDAG is now documented.