Fix a couple of async logic bugs in inciter
1. In LinSysMerger::diagnostics() the reductions, started by signal2host_diag(),
should only be called once the PE has heard from all worker chares, instead of
all the time it hears from a chare. This is now fixed. The symptom for this kind
of problems is a runtime errors with output such as "Reason: mis-matched client
callbacks in reduction messages".

2. Since Conductor::evaluateTime() is only a sufficient condition for fulfilling
all three necessary conditions for wait4report, see conductor.ci, evaluateTime()
should not really mess with wait4report(). This is now placed at the end of
report(), which knows it better, whether waiting for a new report is needed or
not.