Attempt to fix H5Part mac error
Previously, the error (only on mac) was:

dyld: lazy symbol binding failed: Symbol not found: _H5PartOpenFileParallel
  Referenced from: <quinoa>/build/clang/IO/libIO.dylib
  Expected in: flat namespace

dyld: Symbol not found: _H5PartOpenFileParallel
  Referenced from: <quinoa>/build/clang/IO/libIO.dylib
  Expected in: flat namespace

While this appears as a name mangling problem, attempting to fix it as wrapping
the H5Part includes with extern "C", yields compile time errors as mpi.h
attempts to include the OpenMPI C++ API as C-style code. Building H5Part using
the C++ compiler (instead of the C compiler) works, but the above error
persists.

Note that all the above attempts to link against a static H5Part library.

This commit puts in optional static and shared library builds for H5Part,
triggered off of cmake's BUILD_SHARED_LIBS as usual, but that does not help much
either. Now we build a dynamic library on mac (as well as on linux, yielding no
error on linux), and get this even more awful runtime error on mac:

[<hostname>:47315] *** An error occurred in MPI_Info_dup
[<hostname>:47315] *** reported by process [140733618716673,140733193388034]
[<hostname>:47315] *** on communicator MPI_COMM_WORLD
[<hostname>:47315] *** MPI_ERR_INFO: invalid info object
[<hostname>:47315] *** MPI_ERRORS_ARE_FATAL (processes in this communicator will now abort,
[<hostname>:47315] ***    and potentially your MPI job)
^Cmpirun: abort is already in progress...hit ctrl-c again to forcibly terminate